+++ test/component/composer/cases/case49.js
@@ -1,8 +1,8 @@
-export const title       = '49) SVG size undefined';
-export const description = 'component size use the SVG viewBox size';
+export const title       = '49) Size: component';
+export const description = 'component size by css';
 
 export default `
-<g-composer>
+<g-composer style="width: 300px">
    <svg viewBox="0 0 100 100">
      <rect x="0" y="0" width="100" height="100" fill="red"/>
    </svg>