+++ test/component/composer/cases/case50.js
@@ -1,8 +1,8 @@
-export const title       = '50) component size direct';
-export const description = 'by css';
+export const title       = '50) Size: automatic';
+export const description = 'calculated from SVG viewBox';
 
 export default `
-<g-composer style="width: 200px">
+<g-composer>
    <svg viewBox="0 0 100 100">
      <rect x="0" y="0" width="100" height="100" fill="red"/>
    </svg>