+++ CHANGELOG.md
@@ -5,9 +5,12 @@
 - Added `$$()` in the `g-bind` directive to get the current value of the attribute.
 - Improved the code resulting of running `.animateTo()` or `$$.dynamic()` with `transform=rotate()`.
 - Improved Javascript identifier validation in template directives.
-- Improved error description in console messages. 
+- Improved error description in console messages.
+- Fixed a default `width` to 100px when not defined by SVG or component size.
 - Fixed an error with decimal values as `g-for="n of 10.1"`.
 - Fixed an error with load plugin and its internal cache.
+- Removed `--internal-width` CSS variable.
+- Removed support for `hidden` attribute in `g-component`.
 
 ## 0.1.0-alpha.5 (21/11/2023)